+++ src/main/java/snw/rfm/processor/HunterPauseCardProcessor.java
@@ -15,7 +15,9 @@ import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 import org.bukkit.potion.PotionEffect;
 import org.bukkit.potion.PotionEffectType;
+import snw.rfm.RunForMoney;
 import snw.rfm.api.ItemEventListener;
+import snw.rfm.config.GameConfiguration;
 import snw.rfm.config.ItemConfiguration;
 import snw.rfm.game.TeamHolder;
 
@@ -27,9 +29,11 @@ public final class HunterPauseCardProcessor implements ItemEventListener {
             int hpctime = ItemConfiguration.getItemTime("hpc") * 20; // 2022/2/6 避免重复计算
             Bukkit.broadcastMessage(ChatColor.GREEN + player.getName() + " 使用了猎人暂停卡。猎人暂停了 " + (hpctime / 20) + " 秒。");
             for (Player p : th.getHunters()) {
+                p.removePotionEffect(PotionEffectType.SPEED);
                 p.addPotionEffect(new PotionEffect(PotionEffectType.SLOW, hpctime, 255, false));
                 p.addPotionEffect(new PotionEffect(PotionEffectType.JUMP, hpctime, 129, false));
             }
+            Bukkit.getScheduler().runTaskLater(RunForMoney.getInstance(), () -> th.getHunters().forEach(IT -> IT.addPotionEffect(new PotionEffect(PotionEffectType.SPEED, GameConfiguration.getGameTime() * 20, 1, false))), hpctime + 1);
             return true;
         }
         return false;